FIX: Warnings on VC++ compiler

+++ Utilities/NrrdIO/encodingHex.c
@@ -84,7 +84,7 @@ _nrrdEncodingHex_read(Nrrd *nrrd, NrrdIoState *nio) {
   data = nrrd->data;
   nibIdx = 0;
   nibNum = 2*nrrdElementNumber(nrrd)*nrrdElementSize(nrrd);
-  if (nibNum/nrrdElementNumber(nrrd) != 2*nrrdElementSize(nrrd)) {
+  if ((int)(nibNum/nrrdElementNumber(nrrd)) != 2*nrrdElementSize(nrrd)) {
     sprintf(err, "%s: size_t can't hold 2*(#bytes in array)\n", me);
     biffAdd(NRRD, err); return 1;
   }
